AI_CORE/requirements.txt
numpy
librosa
crepe
matplotlib
scipy
tensorflowsounddevice